Low poly blender models creation for chip and dice in progress.

+++ gl3d_engine/src/main/java/com/sadgames/gl3dengine/glrender/scene/objects/Blender3DObject.java
@@ -121,7 +121,7 @@ public class Blender3DObject extends ImportedObject {
         }
 
         verticesArray = new float[verticesList.size() * 3];
-        short vertexPointer = 0;
+        int vertexPointer = 0;
         for (Vector3f vertex : verticesList) {
             verticesArray[vertexPointer++] = vertex.x;
             verticesArray[vertexPointer++] = vertex.y;
